Add reload command
Patch `load.py` (our root module) to reload everything when it loads.
This is for example the case when a user receives an update from
Package Control in the future.

Add `TreeSitterReloadCommand` to mimic touching `load.py`.  Typically
you set in the project settings

```
{
    "build_systems":
    [
        {
            "name": "Reload TreeSitter",
            "target": "tree_sitter_reload"
        },
    ],
    "folders": [
        {
            "path": "."
        },
    ],
    "settings": {
	}
}

```

and then reload on-demand using `ctrl+b`.